add strtol, strcasecmp, strncasecmp, intptr_t to standard library and %o specifier to printf. maybe a few indentation/spacing changes

+++ libc/stddef.h
@@ -2,11 +2,10 @@
 #define __STDDEF_H__
 
 
-#define NULL 			((void *)0)
+#define NULL             ((void *)0)
 
-
-typedef unsigned int 		size_t;
-typedef long long		ptrdiff_t;
+typedef unsigned long    size_t;
+typedef long long        ptrdiff_t;
 
 
 #endif
